Xext: move dbe build into its own subdir
Move the double buffer extension build definitions into
Xext/doublebuffer/meson.build to match the organization of other
extensions.
