+++ Xext/doublebuffer/meson.build
@@ -1,0 +1,10 @@
+srcs_doublebuffer = [
+    'dbe.c',
+    'midbe.c',
+]
+
+libxserver_dbe = static_library('xserver_dbe',
+    srcs_doublebuffer,
+    include_directories: inc,
+    dependencies: common_dep,
+)
