--- dbe/meson.build
@@ -1,10 +1,0 @@
-srcs_dbe = [
-	'dbe.c',
-	'midbe.c',
-]
-
-libxserver_dbe = static_library('xserver_dbe',
-	srcs_dbe,
-	include_directories: inc,
-	dependencies: common_dep,
-)
